web/requirements.txt
flask>=3.0
requests>=2.28
gunicorn>=21.0
python-dotenv>=1.0
